requirements.txt
selenium==4.2.0
python-dotenv==0.20.0
webdriver_manager==3.7.0